add: test_moves
the loss condition was wrong: with the merging mechanic, it can actually be possible to have no empty tiles left while still being able to make a move, we just need to determine which moves are possible based on the current grid state. The test_moves method will clone the current grid and move it to any direction, comparing both grids afterwards. If it changes, the game goes on and the key can be pressed. If not, it's game over.