add: win condition
added wnum var to store the number that has to be reached in order to win the game (2048 by default)
added check_win_condition method that checks if there's a 2048 tile in the board.
refactored new_turn method so it returns one of three possible values: 0, 1, 2, referring to the game's status: in progress, loss and win.