add: pos attribute in piece class
now a position is assigned to each piece in the board.
the user types a position in chess coordinates (C2, for example), and the code checks if the input is typed correctly AND if the position actually corresponds to a piece (using board.find_piece) that belongs to the current team.