add: capture system
while the code is painful to look at, at least it works as intended. For all moving-related functions, I had to add a 'capturing_squares' list to contemplate all positions that has pieces of opposite team.  Also added a piece deleting code at the end of the new_turn method.
the only uncanny thing is being able to capture kings like any other piece, but that will be fixed when checking is added