add: Piece.set_first_move
the first_move attribute would be set to True each time the method to get the possible piece's moves was invoked. But since checking requires the code to see all possible moves for all pieces, it ended up setting to false at the end of the first turn.
to  avoid this, I created a set function to set it to false manually after the end of a turn.