add: end game system!!
finally, there's a way to win the game. or draw, if you dare.
added a method to evaluate if the opposite team has any legal moves at the end of a turn. If it doesn't, it can be either a stalemate or a checkmate, it depends only on if the king is in check.
now that the game can reach an end, the new_turn method also has to return if the game is over or not.
also refactored some of the printing code at the end of the new_turn method so it's easier to understand whats goin on