add: FEN generator, en passant
currently working on updating the fen string after each turn.
in order to do that, I had to add en passant  movement to the game as well as a halfmove/fullmove counter.
now just need to add castling and pretty much all possible moves will be added to the game.